requirements_mws3.txt
# Requirements for running in production. See dev-requirements.txt for
# other requirements.
dj-database-url==0.4.0
django-annoying
django-crispy-forms
django-debug-toolbar
django-email-log
django-extensions
django-filter
django-tables2
django-ucamwebauth
django<1.10
future
requests
whitenoise
djangorestframework
markdown
django-angular


# Testing
mock

pinax-notifications
